Remove GameManager add uml
