+++ space-invaders/core/src/com/samsung/business/spaceinvaders/screens/GameOverScreen.java
@@ -1,21 +1,22 @@
-package com.samsung.business.spaceinvaders;
+package com.samsung.business.spaceinvaders.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
+import com.samsung.business.spaceinvaders.SpaceInvaders;
 import com.samsung.business.spaceinvaders.ui.GameInputMethod;
 import com.samsung.business.spaceinvaders.ui.InputManager;
 import com.samsung.business.spaceinvaders.ui.KeyboardInput;
 
-public class GameWinScreen implements Screen, GameInputMethod {
+public class GameOverScreen implements Screen , GameInputMethod{
     private final SpaceInvaders spaceInvaders;
-    private BitmapFont font;
-    private OrthographicCamera camera;
     private InputManager inputManager;
+    private OrthographicCamera camera;
+    private BitmapFont font;
 
-    public GameWinScreen(SpaceInvaders spaceInvaders) {
+    public GameOverScreen(SpaceInvaders spaceInvaders) {
         font = new BitmapFont();
         Gdx.gl.glClearColor(0, 0, 0, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
@@ -36,9 +37,10 @@ public class GameWinScreen implements Screen, GameInputMethod {
                 break;
         }
         inputManager.setSelectListener(()->{
-            spaceInvaders.setScreen(new GameScreen(spaceInvaders));
+            spaceInvaders.restart();
             dispose();
         });
+
     }
 
     @Override
@@ -50,15 +52,14 @@ public class GameWinScreen implements Screen, GameInputMethod {
     public void render(float delta) {
         camera.update();
         spaceInvaders.batch.setProjectionMatrix(camera.combined);
-
         spaceInvaders.batch.begin();
-        font.draw(spaceInvaders.batch, "GAME WON! " + spaceInvaders.getScore().getValue(), 10, Gdx.graphics.getHeight()/2);
+        font.draw(spaceInvaders.batch, "YOU OVER! " + spaceInvaders.getScore().getValue(), 10, Gdx.graphics.getHeight()/2);
         switch(Gdx.app.getType()) {
             case Android:
-                font.draw(spaceInvaders.batch, "Touch screen to restart", 10, Gdx.graphics.getHeight()/2-50);
+                font.draw(spaceInvaders.batch, "Touch screen to restart", 10, (Gdx.graphics.getHeight()/2)-50);
                 break;
             case Desktop:
-                font.draw(spaceInvaders.batch, "Press enter to restart", 10, Gdx.graphics.getHeight()/2-50);
+                font.draw(spaceInvaders.batch, "Press enter to restart", 10, (Gdx.graphics.getHeight()/2)-50);
 
                 break;
         }
@@ -70,6 +71,7 @@ public class GameWinScreen implements Screen, GameInputMethod {
     @Override
     public void resize(int width, int height) {
         camera.setToOrtho(false, width, height);
+        camera.update();
     }
 
     @Override
