+++ space-invaders/core/src/com/samsung/business/spaceinvaders/screens/GameWinScreen.java
@@ -1,35 +1,30 @@
-package com.samsung.business.spaceinvaders;
+package com.samsung.business.spaceinvaders.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
+import com.samsung.business.spaceinvaders.SpaceInvaders;
 import com.samsung.business.spaceinvaders.ui.GameInputMethod;
 import com.samsung.business.spaceinvaders.ui.InputManager;
 import com.samsung.business.spaceinvaders.ui.KeyboardInput;
 
-public class MainMenuScreen implements Screen , GameInputMethod{
+public class GameWinScreen implements Screen, GameInputMethod {
     private final SpaceInvaders spaceInvaders;
     private BitmapFont font;
+    private OrthographicCamera camera;
     private InputManager inputManager;
 
-    private OrthographicCamera camera;
-
-    private int height = Gdx.graphics.getHeight();
-    private int width = Gdx.graphics.getWidth();
-
-    public MainMenuScreen(SpaceInvaders spaceInvaders) {
+    public GameWinScreen(SpaceInvaders spaceInvaders) {
         font = new BitmapFont();
-        this.spaceInvaders = spaceInvaders;
-
         Gdx.gl.glClearColor(0, 0, 0, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
 
         // tell the camera to update its matrices.
         camera = new OrthographicCamera();
         camera.setToOrtho(false, Gdx.graphics.getWidth(), Gdx.graphics.getHeight());
-
+        this.spaceInvaders = spaceInvaders;
 
         //zaladuj menadzera sterowania
         switch(Gdx.app.getType()) {
@@ -42,8 +37,7 @@ public class MainMenuScreen implements Screen , GameInputMethod{
                 break;
         }
         inputManager.setSelectListener(()->{
-            spaceInvaders.setScreen(new GameScreen(spaceInvaders));
-            dispose();
+            spaceInvaders.restart();
         });
     }
 
@@ -56,29 +50,26 @@ public class MainMenuScreen implements Screen , GameInputMethod{
     public void render(float delta) {
         camera.update();
         spaceInvaders.batch.setProjectionMatrix(camera.combined);
+
         spaceInvaders.batch.begin();
-        font.draw(spaceInvaders.batch, "Space Invaders", 10, height/2);
+        font.draw(spaceInvaders.batch, "GAME WON! " + spaceInvaders.getScore().getValue(), 10, Gdx.graphics.getHeight()/2);
         switch(Gdx.app.getType()) {
             case Android:
-                font.draw(spaceInvaders.batch, "Touch screen to start", 10, height/2-50);
+                font.draw(spaceInvaders.batch, "Touch screen to restart", 10, Gdx.graphics.getHeight()/2-50);
                 break;
             case Desktop:
-                font.draw(spaceInvaders.batch, "Press enter to start", 10, height/2-50);
+                font.draw(spaceInvaders.batch, "Press enter to restart", 10, Gdx.graphics.getHeight()/2-50);
 
                 break;
         }
         spaceInvaders.batch.end();
-
 
         inputManager.update();
     }
 
     @Override
     public void resize(int width, int height) {
-        this.width = width;
-        this.height = height;
         camera.setToOrtho(false, width, height);
-        camera.update();
     }
 
     @Override
