+++ space-invaders/core/src/com/samsung/business/spaceinvaders/ui/ScoreGuiLabel.java
@@ -5,14 +5,20 @@ import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 
 /**
- * Created by lb_lb on 05.11.17.
+ * Created by lb_lb on 18.02.18.
  */
-public class Score {
+
+public class ScoreGuiLabel {
     private final static int MARGIN_TOP = 10;
     private final static int MARGIN_LEFT = 130;
     private int score;
 
-    public Score(int score) {
+    public ScoreGuiLabel() {
+        this.score = 0;
+    }
+
+    public void setScore(int score ){
+
         this.score = score;
     }
 
@@ -21,15 +27,4 @@ public class Score {
         font.draw(batch, String.valueOf(score),  MARGIN_LEFT, Gdx.graphics.getHeight() - MARGIN_TOP);
     }
 
-    public void addScore(int score) {
-        this.score += score;
-    }
-
-    public int getValue() {
-        return score;
-    }
-
-    public void reset() {
-        score = 0;
-    }
 }
